refactor: tweak team modal styles
