+++ web/src/components/TeamModal/TeamModal.module.scss
@@ -1,0 +1,21 @@
+.btn-container {
+  display: flex;
+  justify-content: space-between;
+  margin: 2rem 0 0;
+}
+
+.header {
+  margin: 0 0 2rem;
+}
+
+.input {
+  width: 100%;
+  padding: 0.3rem 0.5rem;
+  display: block;
+  font-size: var(--fontSizeNormal)
+}
+
+.label {
+  margin: 1.5rem 0 0.5rem;
+  display: block;
+}
